[Minuit2] Move MPIProcess header to internals
This class doesn't need to be part of the public interface.
